pyproject.toml
[project]
name = "pegasus"
version = "0.1.0"
description = ""
authors = [
    {name = "0oAstro",email = "79555780+0oAstro@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "pypdf2 (>=3.0.1,<4.0.0)",
    "nltk (>=3.9.1,<4.0.0)",
    "frontend (>=0.0.3,<0.0.4)",
    "sentence-transformers (>=3.3.1,<4.0.0)",
    "beautifulsoup4 (>=4.12.3,<5.0.0)",
    "google-generativeai (>=0.8.3,<0.9.0)",
    "qdrant-client (>=1.12.2,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "streamlit (>=1.41.1,<2.0.0)",
    "watchdog (>=6.0.0,<7.0.0)",
    "groq (>=0.15.0,<0.16.0)",
    "streamlit-lottie (>=0.0.5,<0.0.6)",
    "streamlit-option-menu (>=0.4.0,<0.5.0)",
    "google-cloud-vision (>=3.9.0,<4.0.0)",
    "langchain (>=0.3.14,<0.4.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "pegasus"
package-mode = false